Use strict compiler flags and turn on inlining (#1047)
* Use strict compiler flags and turn on inlining

* Update src/main/scala/com/fulcrumgenomics/fastq/DemuxFastqs.scala

* chore: TransientAttrs.update() to return old or new value

* chore: be more protective in numLongSwitchErrors()

* chore: align => in HapCutToVcf

* chore: add back in the lost but not forgotten G

* chore: manually widen char to int for Java string call

* chore: make small fix to a still-broken test

+++ src/main/scala/com/fulcrumgenomics/alignment/Alignment.scala
@@ -70,7 +70,7 @@ object Cigar {
       require(i > expectedNumberAtPosition, s"Malformed CIGAR string $cigar, found non-digit ${chars(i)} where number expected.")
 
       // Grab the operator (characterToEnum will throw an exception if the operator is invalid)
-      val operator = CigarOperator.characterToEnum(chars(i))
+      val operator = CigarOperator.characterToEnum(chars(i).toInt)
       elements += CigarElem(operator, length)
       i += 1
     }
