+++ src/main/scala/com/fulcrumgenomics/umi/ReviewConsensusVariants.scala
@@ -35,7 +35,7 @@ import com.fulcrumgenomics.umi.ReviewConsensusVariants._
 import com.fulcrumgenomics.util.{Io, Metric}
 import htsjdk.samtools.SamPairUtil.PairOrientation
 import htsjdk.samtools.reference.{ReferenceSequenceFile, ReferenceSequenceFileFactory}
-import htsjdk.samtools.util.{FileExtensions, Interval, IntervalList, Locatable, SamLocusIterator, SequenceUtil}
+import htsjdk.samtools.util._
 import htsjdk.variant.variantcontext.VariantContext
 import htsjdk.variant.vcf.VCFFileReader
 
@@ -276,7 +276,7 @@ class ReviewConsensusVariants
             consensus_read = consensusReadName,
             consensus_insert = toInsertString(rec),
             consensus_call = c.getReadBase.toChar.toUpper,
-            consensus_qual = c.getBaseQuality,
+            consensus_qual = c.getBaseQuality.toInt,
             a = rawCounts.a,
             c = rawCounts.c,
             g = rawCounts.g,
@@ -314,7 +314,7 @@ class ReviewConsensusVariants
     else {
       val list = IntervalList.fromFile(path.toFile).uniqued(false)
       for (i <- list; pos <- i.getStart to i.getEnd) {
-        buffer += Variant(i.getContig, pos, refFile.getSubsequenceAt(i.getContig, pos, pos).getBases()(0).toChar.toUpper, None, None)
+        buffer += Variant(i.getContig, pos, refFile.getSubsequenceAt(i.getContig, pos.toLong, pos.toLong).getBases()(0).toChar.toUpper, None, None)
       }
     }
